+++ src/olmo_core/train/callbacks/slack_notifier.py
@@ -1,14 +1,18 @@
+import logging
 import os
 from dataclasses import dataclass
 from typing import Optional
 
 import requests
 
+from olmo_core.aliases import PathOrStr
 from olmo_core.config import StrEnum
 from olmo_core.distributed.utils import get_rank
 from olmo_core.exceptions import OLMoEnvironmentError
 
 from .callback import Callback
+
+log = logging.getLogger(__name__)
 
 SLACK_WEBHOOK_URL_ENV_VAR = "SLACK_WEBHOOK_URL"
 BEAKER_JOB_ID_ENV_VAR = "BEAKER_JOB_ID"
@@ -77,6 +81,20 @@ class SlackNotifierCallback(Callback):
         if self.notifications == SlackNotificationSetting.all:
             self._post_message("started")
 
+    def post_checkpoint_saved(self, path: PathOrStr):
+        if not self.enabled or get_rank() != 0:
+            return
+
+        if self.notifications == SlackNotificationSetting.all:
+            self._post_message(f"saved a checkpoint to {path}")
+
+    def post_epoch(self):
+        if not self.enabled or get_rank() != 0:
+            return
+
+        if self.notifications == SlackNotificationSetting.all:
+            self._post_message("completed an epoch")
+
     def post_train(self):
         if not self.enabled or get_rank() != 0:
             return
@@ -126,4 +144,7 @@ class SlackNotifierCallback(Callback):
         if BEAKER_JOB_ID_ENV_VAR in os.environ:
             msg = f"{msg}\n*Beaker job:* https://beaker.org/job/{os.environ[BEAKER_JOB_ID_ENV_VAR]}"
 
-        requests.post(webhook_url, json={"text": msg})
+        try:
+            requests.post(webhook_url, json={"text": msg})
+        except Exception as e:
+            log.exception(f"Failed to send Slack notification: {e}")