+++ CHANGELOG.md
@@ -15,13 +15,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Increased NCCL_FASTRAK_PLUGIN_ACCEPT_TIMEOUT_MS from 10 minutes to 30 minutes.
+- Set `fused_linear_cross_entropy_loss accum_dtype` to fp32 in `LMHead`.
+- Increased `NCCL_FASTRAK_PLUGIN_ACCEPT_TIMEOUT_MS` from 10 minutes to 30 minutes.
+- `SlackNotifierCallback` will now notify on checkpoint saved and post epoch events.
+- `BeakerLaunchConfig.launch()` will now send Slack notifications by default when `follow=True` if the env var `SLACK_WEBHOOK_URL` is set.
 
 ### Added
 
 - Adds a custom block that does LayerNorm Scaling
 - Adds the `HalfCos` learning rate scheduler
-
 
 ## [v2.2.0](https://github.com/allenai/OLMo-core/releases/tag/v2.2.0) - 2025-08-26